--- app/src/main/java/com/techcourse/ManualHandlerAdaptor.java
@@ -1,30 +1,0 @@
-package com.techcourse;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapter;
-import com.interface21.webmvc.servlet.view.JspView;
-
-public class ManualHandlerAdaptor implements HandlerAdapter {
-
-    @Override
-    public boolean supports(final Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(
-            final HttpServletRequest request,
-            final HttpServletResponse response,
-            final Object handler
-    ) throws Exception {
-        Controller controller = (Controller) handler;
-        String viewName = controller.execute(request, response);
-        View view = new JspView(viewName);
-        return new ModelAndView(view);
-    }
-}
